correcting master

+++ 15-440Lab3/src/JobEntry.java
@@ -4,12 +4,12 @@ import java.util.HashSet;
 public class JobEntry {
 
 	public String name;
-	public int id;
+	public long id;
 	public String host;
 	public int port;
 	public HashSet<Integer> tasks;
 	
 	public String toString() {
-		return Integer.toString(id) + "  " + name + "  " + host + ":" + Integer.toString(port) + "  " + tasks.toString(); 
+		return Long.toString(id) + "  " + name + "  " + host + ":" + Integer.toString(port) + "  " + tasks.toString(); 
 	}
 }
